Refactor: Move controllers into Notifications namespace
Product and Product kit controllers belong to notifications namespace.
Renamed and wrapped them inside Notifications module/folder for
domain coherence.
