+++ cosmetics-web/app/controllers/responsible_persons/wizard/notifications/product_kit_controller.rb
@@ -1,4 +1,4 @@
-class ResponsiblePersons::Wizard::NotificationProductKitController < SubmitApplicationController
+class ResponsiblePersons::Wizard::Notifications::ProductKitController < SubmitApplicationController
   include Wicked::Wizard
   include CountriesHelper
   include WizardConcern
