+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/_add_product_name_form.html.erb
