+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/_instructions.html.erb
