+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/add_exposure_condition.html.erb
