+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/add_internal_reference.html.erb
