+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/add_product_name.html.erb
