+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/contains_nanomaterials.html.erb
