+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/is_hair_dye.html.erb
