+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product/single_or_multi_component.html.erb
