+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product_kit/is_hair_dye.html.erb
