+++ cosmetics-web/app/views/responsible_persons/wizard/notifications/product_kit/is_mixed.html.erb
