+++ cosmetics-web/config/locales/en.yml
@@ -227,7 +227,7 @@ en:
             selection:
               blank_with_rp_list: "Select a Responsible Person or add a new Responsible Person"
               blank_without_rp_list: "Select add a new Responsible Person"
-        responsible_persons/wizard/notification_product/contains_nanomaterials_form:
+        responsible_persons/wizard/notifications/product/contains_nanomaterials_form:
           attributes:
             contains_nanomaterials:
               inclusion: "Select yes if the product is a multi-item kit, no if its single item"
@@ -236,7 +236,7 @@ en:
               not_an_integer: "Enter a number for how many nanomaterials"
               greater_than: "Enter a number for how many nanomaterials"
               less_than_or_equal_to: "Maximum nanomaterials count is 10. More can be added later"
-        responsible_persons/wizard/notification_product/single_or_multi_component_form:
+        responsible_persons/wizard/notifications/product/single_or_multi_component_form:
           attributes:
             single_or_multi_component:
               inclusion: "Select yes if the product is a multi-item kit, no if its single item"
